server/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
openai
google-generativeai
requests
pgvector
pyhumps
httpx
